Get Relay service running again
Building the web site to another folder other than directly underneath it breaks it at runtime. I had to prevent ReadOnlySourceTree from propagating to the web project to fix this.

At this point, I still get a 403 Forbidden, but it's because it's configured to use Azure storage but with an invalidated key.